+++ wemake_python_styleguide/visitors/ast/classes.py
@@ -9,36 +9,22 @@ from typing_extensions import final
 from wemake_python_styleguide import constants, types
 from wemake_python_styleguide.compat.aliases import AssignNodes
 from wemake_python_styleguide.compat.functions import get_assign_targets
-from wemake_python_styleguide.logic.classes import is_forbidden_super_class
-from wemake_python_styleguide.logic.functions import get_all_arguments
+from wemake_python_styleguide.logic import classes, functions
 from wemake_python_styleguide.logic.nodes import (
     get_context,
     is_contained,
     is_doc_string,
 )
-from wemake_python_styleguide.types import AnyAssign
-from wemake_python_styleguide.violations.best_practices import (
-    BadMagicMethodViolation,
-    BaseExceptionSubclassViolation,
-    BuiltinSubclassViolation,
-    IncorrectBaseClassViolation,
-    IncorrectClassBodyContentViolation,
-    IncorrectSlotsViolation,
-    MethodWithoutArgumentsViolation,
-    ShadowedClassAttributeViolation,
-    StaticMethodViolation,
-    YieldInsideInitViolation,
-)
+from wemake_python_styleguide.violations import best_practices as bp
 from wemake_python_styleguide.violations.consistency import (
     ObjectInBaseClassesListViolation,
     RequiredBaseClassViolation,
 )
-from wemake_python_styleguide.visitors.base import BaseNodeVisitor
-from wemake_python_styleguide.visitors.decorators import alias
+from wemake_python_styleguide.visitors import base, decorators
 
 
 @final
-class WrongClassVisitor(BaseNodeVisitor):
+class WrongClassVisitor(base.BaseNodeVisitor):
     """
     This class is responsible for restricting some ``class`` anti-patterns.
 
@@ -70,19 +56,19 @@ class WrongClassVisitor(BaseNodeVisitor):
     def _check_base_classes(self, node: ast.ClassDef) -> None:
         for base_name in node.bases:
             if not isinstance(base_name, self._allowed_base_classes_nodes):
-                self.add_violation(IncorrectBaseClassViolation(node))
+                self.add_violation(bp.IncorrectBaseClassViolation(node))
                 continue
 
             id_attr = getattr(base_name, 'id', None)
             if id_attr == 'BaseException':
-                self.add_violation(BaseExceptionSubclassViolation(node))
+                self.add_violation(bp.BaseExceptionSubclassViolation(node))
             elif id_attr == 'object' and len(node.bases) >= 2:
                 self.add_violation(
                     ObjectInBaseClassesListViolation(node, text=id_attr),
                 )
-            elif is_forbidden_super_class(id_attr):
+            elif classes.is_forbidden_super_class(id_attr):
                 self.add_violation(
-                    BuiltinSubclassViolation(node, text=id_attr),
+                    bp.BuiltinSubclassViolation(node, text=id_attr),
                 )
 
     def _check_wrong_body_nodes(self, node: ast.ClassDef) -> None:
@@ -91,7 +77,7 @@ class WrongClassVisitor(BaseNodeVisitor):
                 continue
             if is_doc_string(sub_node):
                 continue
-            self.add_violation(IncorrectClassBodyContentViolation(sub_node))
+            self.add_violation(bp.IncorrectClassBodyContentViolation(sub_node))
 
     def visit_ClassDef(self, node: ast.ClassDef) -> None:
         """
@@ -111,11 +97,11 @@ class WrongClassVisitor(BaseNodeVisitor):
 
 
 @final
-@alias('visit_any_function', (
+@decorators.alias('visit_any_function', (
     'visit_FunctionDef',
     'visit_AsyncFunctionDef',
 ))
-class WrongMethodVisitor(BaseNodeVisitor):
+class WrongMethodVisitor(base.BaseNodeVisitor):
     """Visits functions, but treats them as methods."""
 
     _staticmethod_names: ClassVar[FrozenSet[str]] = frozenset((
@@ -146,33 +132,35 @@ class WrongMethodVisitor(BaseNodeVisitor):
         for decorator in node.decorator_list:
             decorator_name = getattr(decorator, 'id', None)
             if decorator_name in self._staticmethod_names:
-                self.add_violation(StaticMethodViolation(node))
+                self.add_violation(bp.StaticMethodViolation(node))
 
     def _check_bound_methods(self, node: types.AnyFunctionDef) -> None:
         node_context = get_context(node)
         if not isinstance(node_context, ast.ClassDef):
             return
 
-        if not get_all_arguments(node):
+        if not functions.get_all_arguments(node):
             self.add_violation(
-                MethodWithoutArgumentsViolation(node, text=node.name),
+                bp.MethodWithoutArgumentsViolation(node, text=node.name),
             )
 
         if node.name in constants.MAGIC_METHODS_BLACKLIST:
-            self.add_violation(BadMagicMethodViolation(node, text=node.name))
+            self.add_violation(
+                bp.BadMagicMethodViolation(node, text=node.name),
+            )
 
     def _check_method_contents(self, node: types.AnyFunctionDef) -> None:
         if node.name == constants.INIT:
             if is_contained(node, self._not_appropriate_for_init):
-                self.add_violation(YieldInsideInitViolation(node))
+                self.add_violation(bp.YieldInsideInitViolation(node))
 
 
 @final
-@alias('visit_any_assign', (
+@decorators.alias('visit_any_assign', (
     'visit_Assign',
     'visit_AnnAssign',
 ))
-class WrongSlotsVisitor(BaseNodeVisitor):
+class WrongSlotsVisitor(base.BaseNodeVisitor):
     """Visits class attributes."""
 
     _blacklisted_slots_nodes: ClassVar[types.AnyNodes] = (
@@ -181,7 +169,7 @@ class WrongSlotsVisitor(BaseNodeVisitor):
         ast.Set,
     )
 
-    def visit_any_assign(self, node: AnyAssign) -> None:
+    def visit_any_assign(self, node: types.AnyAssign) -> None:
         """
         Checks all assigns that have correct context.
 
@@ -192,7 +180,7 @@ class WrongSlotsVisitor(BaseNodeVisitor):
         self._check_slots(node)
         self.generic_visit(node)
 
-    def _contains_slots_assign(self, node: AnyAssign) -> bool:
+    def _contains_slots_assign(self, node: types.AnyAssign) -> bool:
         targets = get_assign_targets(node)
 
         for target in targets:
@@ -202,22 +190,22 @@ class WrongSlotsVisitor(BaseNodeVisitor):
 
     def _count_slots_items(
         self,
-        node: AnyAssign,
+        node: types.AnyAssign,
         elements: ast.Tuple,
     ) -> None:
         fields: List[str] = []
         for tuple_item in elements.elts:
             if not isinstance(tuple_item, ast.Str):
-                self.add_violation(IncorrectSlotsViolation(node))
+                self.add_violation(bp.IncorrectSlotsViolation(node))
                 return
             fields.append(tuple_item.s)
 
         for _, counter in Counter(fields).items():
             if counter > 1:
-                self.add_violation(IncorrectSlotsViolation(node))
+                self.add_violation(bp.IncorrectSlotsViolation(node))
                 return
 
-    def _check_slots(self, node: AnyAssign) -> None:
+    def _check_slots(self, node: types.AnyAssign) -> None:
         if not isinstance(get_context(node), ast.ClassDef):
             return
 
@@ -225,7 +213,7 @@ class WrongSlotsVisitor(BaseNodeVisitor):
             return
 
         if isinstance(node.value, self._blacklisted_slots_nodes):
-            self.add_violation(IncorrectSlotsViolation(node))
+            self.add_violation(bp.IncorrectSlotsViolation(node))
             return
 
         if isinstance(node.value, ast.Tuple):
@@ -233,11 +221,14 @@ class WrongSlotsVisitor(BaseNodeVisitor):
 
 
 @final
-class ClassAttributeVisitor(BaseNodeVisitor):
+class ClassAttributeVisitor(base.BaseNodeVisitor):
     """Finds incorrect class attributes."""
 
     # TODO: can be moved to logic, if is used anywhere else
-    def _flat_assign_names(self, nodes: List[AnyAssign]) -> Container[str]:
+    def _flat_assign_names(
+        self,
+        nodes: List[types.AnyAssign],
+    ) -> Container[str]:
         flat_assigns = []
         for attribute in nodes:
             targets = get_assign_targets(attribute)
@@ -250,7 +241,7 @@ class ClassAttributeVisitor(BaseNodeVisitor):
     def _get_attributes(
         self,
         node: ast.ClassDef,
-    ) -> Tuple[List[AnyAssign], List[ast.Attribute]]:
+    ) -> Tuple[List[types.AnyAssign], List[ast.Attribute]]:
         class_attributes = []
         instance_attributes = []
 
@@ -270,7 +261,7 @@ class ClassAttributeVisitor(BaseNodeVisitor):
         for instance_attr in instance_attributes:
             if instance_attr.attr in class_attribute_names:
                 self.add_violation(
-                    ShadowedClassAttributeViolation(
+                    bp.ShadowedClassAttributeViolation(
                         instance_attr, text=instance_attr.attr,
                     ),
                 )