+++ CHANGELOG.md
@@ -30,6 +30,9 @@ consistency, complexity, and general feel from your code.
 - Forbids to have class variables that are shadowed by instance variables
 - Forbids to use ternary expressions inside `if` conditions
 - Forces to use ternary instead of `... and ... or ...` expression
+- Forces to use `a > b > c` instead of `a > b and b > c`
+- Forces to use `c < b < a` instead of `a > b > c`
+- Forbids to use explicit `in []` and `in ()`, use sets or variables instead
 
 ### Bugfixes
 
@@ -43,6 +46,7 @@ consistency, complexity, and general feel from your code.
 - Renames `Redundant` to `Useless`
 - Renames `Comparison` to `Compare`
 - Renames `WrongConditionalViolation` to `ConstantConditionViolation`
+- Renames `ComplexDefaultValuesViolation` to `ComplexDefaultValueViolation`
 - Refactors `UselessOperatorsVisitor`
 - Adds `compat/` package, getting ready for `python3.8`
 - Adds `Makefile`
@@ -67,7 +71,7 @@ and lots of bug fixes.
 - Fixes problem with missing `_allowed_left_nodes`
 - Fixes problem false positive for `Z121` when using `_` for unused var names
 - Fixes false positive for negative number in default values
-- Fixes error text for `ComplexDefaultValuesViolation`
+- Fixes error text for `ComplexDefaultValueViolation`
 - Fixes problem with false positive for `Z459`
   when a default value is an `Ellipsis`